+++ angular-frontend/src/app/omnai-datasource/omnai-scope-server/live-data.service.ts
@@ -6,6 +6,11 @@ import { catchError, Observable, of, Subject, switchMap, takeUntil, timer } from
 import { map, filter } from 'rxjs/operators';
 import { BackendPortService } from './backend-port.service';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import { take } from 'rxjs/operators';
+import { SaveDataLocallyModalComponent } from '../../save-data-locally-modal/save-data-locally-modal.component';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { DownloadProgressComponent } from './downloadProgress.component';
 
 interface DeviceInformation {
   UUID: string;
@@ -26,17 +31,34 @@ interface DeviceOverview {
   }[];
 }
 
+// WSMessageInterfaces 
+
+interface DataMessage {
+  type: 'data';
+  devices: string[];
+  data: { timestamp: number, value: number[] }[];
+}
+
+interface FileReadyMessage {
+  type: 'file-ready',
+  url: string
+}
+
+type WSMessage = DataMessage | FileReadyMessage;
+
 @Injectable({
   providedIn: 'root'
 })
 export class OmnAIScopeDataService implements DataSource {
 
   private readonly destroyRef = inject(DestroyRef);
+  private readonly dialog = inject(MatDialog);
 
-  constructor() {
+  constructor(private snackBar: MatSnackBar) {
     this.setupDevicePolling();
   }
   private socket: WebSocket | null = null;
+
 
   readonly isConnected = signal<boolean>(false);
   readonly devices = signal<DeviceInformation[]>([]);
@@ -53,6 +75,7 @@ export class OmnAIScopeDataService implements DataSource {
   });
 
   readonly #httpClient = inject(HttpClient);
+  private readonly fileReady$ = new Subject<FileReadyMessage>();
 
   private setupDevicePolling(): void {
     const pollInterval_ms = 15 * 1000;
@@ -128,34 +151,40 @@ export class OmnAIScopeDataService implements DataSource {
 
     let ignoreCounter = 0;
     this.socket.addEventListener('message', (event) => {
-      if (ignoreCounter < 2) {
-        // ignore first messages sometimes 
-        ignoreCounter++;
-        return;
-      }
+      const raw = (event.data as string).trim();
 
-      let parsedMessage: any;
+      if (!raw.startsWith('{')) return;
+
+      let parsedMessage: WSMessage;
       try {
         parsedMessage = JSON.parse(event.data);
       } catch {
-        parsedMessage = event.data;
+        console.warn("Wrong JSON format", event.data);
+        return;
       }
 
-      if (this.isOmnAIDataMessage(parsedMessage)) {
-        this.data.update(records => {
-          parsedMessage.devices.forEach((uuid: string, index: number) => {
-            const existingData = records[uuid] ?? [];
-            const newDataPoints = parsedMessage.data.map((point: any) => ({
-              timestamp: point.timestamp,
-              value: point.value[index],
-            }));
-            records[uuid] = existingData.concat(newDataPoints);
-          });
+      switch (parsedMessage.type) {
+        case 'data':
+          this.data.update(records => {
+            parsedMessage.devices.forEach((uuid: string, index: number) => {
+              const existingData = records[uuid] ?? [];
+              const newDataPoints = parsedMessage.data.map((point: any) => ({
+                timestamp: point.timestamp,
+                value: point.value[index],
+              }));
+              records[uuid] = existingData.concat(newDataPoints);
+            });
 
-          return { ...records };
-        });
-      } else {
-        console.warn('Unbekanntes Nachrichtenformat:', parsedMessage);
+            return { ...records };
+          });
+          break;
+        case 'file-ready':
+          console.log("File was ready");
+          this.fileReady$.next(parsedMessage);
+          break;
+        default:
+          console.warn("Not known message format");
+          console.log(parsedMessage);
       }
     });
 
@@ -172,37 +201,60 @@ export class OmnAIScopeDataService implements DataSource {
 
   // close websocket connection
   disconnect(): void {
-    if (this.socket?.readyState === WebSocket.OPEN) this.socket.close();
+    if (this.socket) {
+      let stopMessage = {
+        type: 'stop'
+      }
+      console.log(JSON.stringify(stopMessage));
+      this.socket.send(JSON.stringify(stopMessage));
+    }
   }
 
   clearData(): void {
-    this.data.set({});
+    if (this.socket) {
+      this.socket.close();
+      this.socket = null;
+      this.isConnected.set(false);
+      this.data.set({});
+    }
   }
 
-  // Typprüfung für OmnAI-Daten-Nachrichten
-  private isOmnAIDataMessage(message: any): boolean {
-    if (typeof message !== 'object' || message === null) return false;
+  save(): void {
+    const dialogRef = this.dialog.open(SaveDataLocallyModalComponent, {
+      width: '60vw'
+    });
+    dialogRef.afterClosed().pipe(filter(Boolean)).subscribe(({ dir, fileName }) => {
+      let serverpath = `/download/${fileName}`;
+      const saveMessage = {
+        type: `save`,
+        uuids: this.devices().map(device => device.UUID),
+        path: `${fileName}`,
+        format: `csv`
+      }
+      const progressRef = this.dialog.open(DownloadProgressComponent, {
+        disableClose: true
+      });
+      this.fileReady$
+        .pipe(
+          filter(m => m.url === serverpath),
+          take(1)
+        )
+        .subscribe(() => {
+          window.electronAPI?.downloadFile(serverpath, dir, fileName)
+            .then(() => {
+              this.snackBar.open('File sucessfully saved ✓', '', { duration: 4000 });
+            })
+            .catch(() => {
+              this.snackBar.open('Saving error', '', { duration: 4000 });
+            })
+            .finally(() => progressRef.close());
+        });
+      this.socket?.send(JSON.stringify(saveMessage));
+      console.log("dialog closed");
+    });
+  }
 
-    if (!('devices' in message) || !('data' in message)) return false;
-    if (
-      !Array.isArray(message.devices) ||
-      !message.devices.every((d: unknown) => typeof d === 'string')
-    ) {
-      return false;
-    }
-
-    if (
-      !Array.isArray(message.data) ||
-      !message.data.every(
-        (entry: any) =>
-          typeof entry.timestamp === 'number' &&
-          Array.isArray(entry.value) &&
-          entry.value.every((v: unknown) => typeof v === 'number'),
-      )
-    ) {
-      return false;
-    }
-
-    return true;
+  record(): void {
+    console.log('Start recording OmnAI data ...');
   }
 }