+++ angular-frontend/src/app/toolbar/start-data-button.component.ts
@@ -1,10 +1,11 @@
 
 import { Component, inject } from '@angular/core';
 import { MatDialog, MatDialogModule } from '@angular/material/dialog';
-import { DataSourceSelectionService } from './data-source-selection.service';
-import { SourceSelectModalComponent } from './source-select-modal.component';
+import { DataSourceSelectionService } from '../source-selection/data-source-selection.service';
+import { SourceSelectModalComponent } from '../source-selection/source-select-modal.component';
 import { MatIconModule } from '@angular/material/icon';
 import { AdvancedModeService } from '../advanced-mode/advanced-mode.service';
+import { ToolbarState, ToolbarStateManagerService } from './toolbarStateManager.service';
 
 @Component({
     selector: 'app-start-data-button',
@@ -12,7 +13,7 @@ import { AdvancedModeService } from '../advanced-mode/advanced-mode.service';
     imports: [MatDialogModule, MatIconModule],
     template: `
         <button mat-icon-button (click)="toggleStartButton()" aria-label="Start Data" id="start-button">
-        <mat-icon>{{ measurementIsStarted ? 'stop' : 'play_arrow' }}</mat-icon>
+        <mat-icon>{{ (this.toolbarState.getState() === ToolbarState.STARTED ) ? 'stop' : 'play_arrow' }}</mat-icon>
         </button>
     `,
     styles: `button { display: flex; padding: .3em }`,
@@ -21,8 +22,8 @@ export class StartDataButtonComponent {
     private readonly dialog = inject(MatDialog);
     private readonly datasource = inject(DataSourceSelectionService);
     private readonly advancedMode = inject(AdvancedModeService);
-
-    protected measurementIsStarted: boolean = false;
+    readonly toolbarState = inject(ToolbarStateManagerService);
+    ToolbarState = ToolbarState;
 
     clearAllData(): void {
         this.datasource.availableSources().forEach((source) => {
@@ -38,8 +39,10 @@ export class StartDataButtonComponent {
             });
             dialogRef.afterClosed().subscribe(() => {
                 if (this.datasource.hasSelection()) {
+                    this.toolbarState.setState(ToolbarState.STARTED);
                     this.datasource.currentSource()?.connect();
                 }
+                else this.toolbarState.setState(ToolbarState.IDLE);
             });
             return;
         }
@@ -48,6 +51,7 @@ export class StartDataButtonComponent {
             if (OmnAIScope) {
                 this.datasource.selectSource(OmnAIScope);
                 OmnAIScope.connect();
+                this.toolbarState.setState(ToolbarState.STARTED);
             }
         }
     }
@@ -57,7 +61,15 @@ export class StartDataButtonComponent {
     }
 
     toggleStartButton(): void {
-        this.measurementIsStarted ? this.stopMeasurement() : this.openModal();
-        this.measurementIsStarted = !this.measurementIsStarted;
+        switch (this.toolbarState.getState()) {
+            case ToolbarState.IDLE: // fall through on purpose  
+            case ToolbarState.STOPPED:
+                this.openModal();
+                break;
+            case ToolbarState.STARTED:
+                this.stopMeasurement();
+                this.toolbarState.setState(ToolbarState.STOPPED);
+                break;
+        }
     }
 }