+++ electron/src/preload.ts
@@ -1,7 +1,10 @@
-import { contextBridge, ipcRenderer } from "electron";
+import { contextBridge, ipcRenderer, webUtils } from "electron";
 
 // this adds a new object to the window object. Only the set functions are available instead of random node functions 
 // secure bridge between renderer and main process 
 contextBridge.exposeInMainWorld('electronAPI', {
-    getOmnAIScopeBackendPort: (): Promise<number> => ipcRenderer.invoke('get-omnaiscope-backend-port')
+    getOmnAIScopeBackendPort: (): Promise<number> => ipcRenderer.invoke('get-omnaiscope-backend-port'),
+    downloadFile: (serverpath: string, dir: string, fileName: string) => ipcRenderer.invoke('download-file', { serverpath, dir, fileName }),
+    getAbsolutePath: (file: File): string => webUtils.getPathForFile(file),
+    saveFile: (data: string, folderPath: string, fileName: string) => ipcRenderer.invoke('save-file', { data, folderPath, fileName })
 });